For EXP-2012: Move tests for persistence into their own file
